==> seqFISH_analysis/spatial_statistics/02_compute_cell_centers.ipynb <==
# Compute cell centers from cell masks

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
from tqdm import tqdm, trange
import os
import sys
import dask
import dask.array as da
import dask.dataframe as dd
from dask.diagnostics import ProgressBar
from skimage.io import imread
import tifffile as tf

# directories and inputs

# # folder paths for PKL files with RNA positions
# dotPaths = []
# dotPaths.append(r"Y:\coskun-lab\Zhou\4_HCR\20210614_hchCulture\00_registered_images\detected_dots")
# dotPaths.append(r"Y:\coskun-lab\Zhou\4_HCR\20210507_cytokines\uc\00_registered_images\detected_dot_positions")
# dotPaths.append(r"Y:\coskun-lab\Zhou\4_HCR\20210324_11gene\BM\registered images\detected_dots_positions_v2")
# dotPaths.append(r"Y:\coskun-lab\Zhou\4_HCR\20210324_11gene\UC\00_registered_images\detected_dots_positions")

# folder with mask paths for different tissue sources
maskPaths = []
maskPaths.append(r"..\images\HBM\registered")
maskPaths.append(r"..\images\HUC\registered")
maskPaths.append(r"..\images\HCH\registered")
cwd = os.getcwd()

exportPath = r"..\results\spatial_statistics"

idCols = ['CellY', 'CellX', 'FOV', 'CellLabel', 'TissueSource']

# For each FOV, read cell masks and compute cell centers

dfCenters = pd.DataFrame(columns = idCols)

lazy_read = dask.delayed(imread)

for ii, tissuePath in enumerate(maskPaths): # each tissue source
    
    os.chdir(tissuePath)
    
    tissueSource = tissuePath.split(os.sep)[-3]
    tissueSource = tissueSource.split('_')[-1]
    if 'Culture' in tissueSource:
        tissueSource = tissueSource.replace('Culture', '')
        
    tissueSource = tissueSource.upper() # all caps
    
    # find all FOVs
    fovs = os.listdir()
    
    # preserve cell count across FOVs.  Start with max of previous
    dfSub = dfCenters.loc[dfCenters['TissueSource'] == tissueSource]
    if dfSub.size == 0:
        cellCount = 0 # no cells recorded yet
        
    else:
        cellCount = dfSub['CellLabel'].max()
    
    for jj, fov in enumerate(tqdm(fovs)): # each FOV
        
        os.chdir(fov)
        
        # find all cell masks for this FOV
        cellFiles = [f for f in os.listdir() if f.endswith('.tif') and 'cell' in f and 'Mask' in f]
        if len(cellFiles) == 0: # empty, no cells
            os.chdir('..') # go back to main folder
            continue # next iter
        
        # get image dimensions
        img = tf.TiffFile(cellFiles[0])
        dimY, dimX = img.pages[0].shape
        dtype = img.pages[0].dtype
        img.close()
        
        for kk, cellFile in enumerate(cellFiles): # each cell mask TIF
            
            # cellLabel = cellFile.split('_')[0].replace('cell', '')
            # cellLabel = int(cellLabel)
            cellCount += 1
            
            # read cell mask
            mask = lazy_read(cellFile)
            mask = da.from_delayed(mask, shape = [dimY, dimX], dtype = dtype)
            
            # calculate cell center of mass
            y, x = np.where(mask > 0)
            y.compute_chunk_sizes()
            x.compute_chunk_sizes()
                        
            # append to dataframe
            dfSub = pd.DataFrame()
            dfSub['TissueSource'] = [tissueSource]
            dfSub['FOV'] = [int(fov)]
            dfSub['CellLabel'] = [cellCount]
            dfSub['CellY'] = [np.mean(y).compute()]
            dfSub['CellX'] = [np.mean(x).compute()]
            
            # check for empty mask
            if dfSub['CellY'].isna().any():
                print()
                print('Mask empty:')
                print(tissuePath)
                print(fov)
                print(cellFile)
                print(cellCount)
            
            dfCenters = pd.concat([dfCenters, dfSub])    
        
        os.chdir('..') # go back to main folder
        
    
dfCenters = dfCenters.reset_index(drop = True)

dfCenters

dfCenters.groupby(['TissueSource']).max()

# Export cell centers dataframe

os.chdir(exportPath)

dfCenters.to_pickle('02_cell_centers.pkl')
