==> seqFISH_analysis/spatial_statistics/03_compute_dist_centers.ipynb <==
# Compute distance to cell centers for all RNA spots

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
from tqdm import tqdm, trange
import os
import sys
import dask
import dask.array as da
import dask.dataframe as dd
from dask.diagnostics import ProgressBar
from skimage.io import imread
import tifffile as tf
import torch

# directories and inputs

dataPath = r"..\results\spatial_statistics"

idCols = ['Y', 'X', 'CellY', 'CellX', 'FOV', 'CellLabel', 'TissueSource']

# Read RNA spots data and cell centers.  Merge dataframes

os.chdir(dataPath)

dfSpots = pd.read_pickle('01_all_RNA_positions.pkl')
dfCenters = pd.read_pickle('02_cell_centers.pkl')

# merge dataframes based on cell labels
dfAll = dfSpots.merge(dfCenters, how = 'outer')

# drop rows (pixels) if any nan.  These may be RNA spots w/o corresponding masks or masks w/o corresponding RNA spots
# markers = dfAll.drop(columns = idCols).columns.tolist()
dfAll.dropna(axis = 0, how = 'any', inplace = True)

dfAll

# Compute distance to cell center for every spot

# distCenters = []
# for ii in trange(dfAll.shape[0]): # each row/spot
    
#     row = dfAll.iloc[ii, :]
#     v1 = row[['Y', 'X']].values
#     v2 = row[['CellY', 'CellX']].values
#     dist = np.linalg.norm( - )
    
#     distCenters.append(dist)

cuda = torch.device('cuda')     # Default CUDA device
v1 = torch.tensor(dfAll[['Y', 'X']].values, device = cuda)
v2 = torch.tensor(dfAll[['CellY', 'CellX']].values, device = cuda)

# use GPU for faster calculations
distCenters = torch.square(v1 - v2)
distCenters = torch.sum(distCenters, axis = 1)
distCenters = torch.sqrt(distCenters)
distCenters = distCenters.cpu() # bring back to CPU

# append to dataframe
dfAll['DistCenter'] = distCenters

# relabel tissue sources
dfAll['TissueSource'].replace(to_replace = {'BM': 'HBM', 'UC': 'HUC'}, inplace = True)

dfAll

# export dataframe

os.chdir(dataPath)

dfAll.to_pickle('03_spots_wt_distCenters.pkl')
